requirements.txt
pydantic
python-pptx
requests
flask
python-dotenv
google-genai
langgraph
langchain
langchain-google-genai
APScheduler
